fix: align output e2e tests with rendered dom
